#871-delete-snapshot - delete drives snapshots

+++ src/app/vm/vm-sidebar/storage-detail/volumes/snapshot/snapshot-modal.container.ts
@@ -1,14 +1,7 @@
-import {
-  Component,
-  Inject,
-  OnInit
-} from '@angular/core';
+import { Component, Inject, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { State } from '../../../../../reducers/index';
-import {
-  MAT_DIALOG_DATA,
-  MatDialogRef
-} from '@angular/material';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
 import { Volume } from '../../../../../shared/models/volume.model';
 import { Snapshot } from '../../../../../shared/models/snapshot.model';
 import { WithUnsubscribe } from '../../../../../utils/mixins/with-unsubscribe';
@@ -16,6 +9,7 @@ import { WithUnsubscribe } from '../../../../../utils/mixins/with-unsubscribe';
 import * as volumeActions from '../../../../../reducers/volumes/redux/volumes.actions';
 import * as snapshotActions from '../../../../../reducers/snapshots/redux/snapshot.actions';
 import * as fromVolumes from '../../../../../reducers/volumes/redux/volumes.reducers';
+import { DialogService } from '../../../../../dialog/dialog-service/dialog.service';
 
 @Component({
   selector: 'cs-snapshot-modal-container',
@@ -35,6 +29,7 @@ export class SnapshotModalContainerComponent extends WithUnsubscribe() implement
     @Inject(MAT_DIALOG_DATA) data,
     public dialogRef: MatDialogRef<SnapshotModalContainerComponent>,
     private store: Store<State>,
+    private dialogService: DialogService
   ) {
     super();
     this.store.dispatch(new volumeActions.LoadSelectedVolume(data.volumeId));
@@ -54,6 +49,11 @@ export class SnapshotModalContainerComponent extends WithUnsubscribe() implement
   }
 
   public snapshotDeleted(snapshot: Snapshot) {
-    this.store.dispatch(new snapshotActions.DeleteSnapshot(snapshot));
+    this.dialogService.confirm({ message: 'DIALOG_MESSAGES.SNAPSHOT.CONFIRM_DELETION' })
+      .onErrorResumeNext()
+      .filter(res => Boolean(res))
+      .subscribe(() => {
+        this.store.dispatch(new snapshotActions.DeleteSnapshot(snapshot));
+      });
   }
 }
